Create a RockModel parent to inherit from
