+++ api/src/data/content-channels/index.js
@@ -20,9 +20,10 @@ export const resolver = {
   },
   ContentChannel: {
     id: ({ id }, _, $, { parentType }) => createGlobalId(id, parentType.name),
-    childrenContentItems: ({ id }, pagination, { models }) => ({
-      cursor: models.ContentItem.byContentChannelId(id),
-      pagination,
-    }),
+    childrenContentItems: ({ id }, input, { models }) =>
+      models.ContentItem.paginate({
+        cursor: models.ContentItem.byContentChannelId(id),
+        input,
+      }),
   },
 };
